+++ packages/rust/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod config;
+
 use clap::{Parser, Subcommand};
 
 #[derive(Parser, Debug)]